# Fix batched matmul for integral numbers