Close #6715.

(UPDATED)This PR fixes batched matmul for integral numbers by implementing custom kernels on top of non-batched one.