@emcastillo 2022-05-17
Does this work? there is an specific kernel used for integer matrix multiplication so cublas shouldn't be used.
If this is working there is something wrong in the cublas/kernel selection 😂