@takagi 2022-05-17
Yes, but I think the kernel doesn't work as batched one. Implementing another for batched gemm is the right way.

https://github.com/cupy/cupy/blob/8968a3e04cfa84bd6d22f33c63e65b9fbd24d041/cupy/_core/_routines_linalg.pyx#L103