@takagi 2022-05-17
Ah, we can compute batched gemm almost with the kernel in parallel by blocks of threads. I'm going to implement it soon. Thanks for your pointing!