@takagi 2022-05-19
Seems that the CI error is caused by SciPy 1.8.1. (Can be reproduced locally)