@takagi 2022-05-20
Comparison between master and this PR, with int32 matrix batches. I get 10~20%  improvement in efficiency with this PR. BTW, as touched in #6703, I also observed that batched matmul on the same dtype size (float32) matrices runs ~30% faster with cublas routines. Maybe they're well tuned for each Nvidia device.

shape|master[ms]|#6725[ms]|numpy[ms]
---|---|---|---
(16, 256, 256)| 0.15575040131807327| 0.14878720194101336| 183.2779342651367
(32, 256, 256)| 0.24320000112056736| 0.21473280340433124| 453.9682891845704
(64, 256, 256)| 0.43950079977512363| 0.3778559982776642| 716.5655822753906
(128, 256, 256)| 0.8155135989189148| 0.6744063913822175| 2409.152319335937
(256, 256, 256)| 1.5748095870018006| 1.2906496286392215| 4853.44130859375