@emcastillo 2022-05-20
The kernel that we are using could be much faster, there is a PR open trying to tune it, but seems difficult.