@emcastillo 2022-05-24
Seems like hip doesn't like this kernel