@takagi 2022-06-06
Seems that hiprtc doesn't permit calling `__global__` function from another `__global__`.