@leofang 2022-06-06
> Seems that hiprtc doesn't permit calling `__global__` function from another `__global__`.

Right, on the CUDA side it's called dynamic parallelism l, which requires separate compilation, but back then it was not supported on HIP https://github.com/cupy/cupy/blob/3ffd5f359246d9afd9fd51be3b8bd191d777ef81/cupy/cuda/compiler.py#L835-L836 not sure about now.