@takagi 2022-06-07
> Right, on the CUDA side it's called dynamic parallelism l, which requires separate compilation, but back then it was not supported on HIP

Thanks! Got it. Seems that the situation is still not changed.